Add sharedElementTypeSelectionMock
